# Improve performance substantially by parameterizing types