--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -9,8 +9,9 @@ jobs:
     strategy:
       matrix:
         version:
-          - '1.1'
           - '1.5'
+          - '1.6'
+          - 'nightly'
         os:
           - ubuntu-latest
           - macOS-latest