--- a/Project.toml
+++ b/Project.toml
@@ -1,7 +1,7 @@
 name = "Yields"
 uuid = "d7e99b2f-e7f3-4d9e-9f01-2338fc023ad3"
 authors = ["Alec Loudenback <alecloudenback@gmail.com> and contributors"]
-version = "0.7.1"
+version = "0.8.0"
 
 [deps]
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"