--- a/Project.toml
+++ b/Project.toml
@@ -11,7 +11,7 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 [compat]
 ForwardDiff = "^0.10"
 Interpolations = "^0.12.4, 0.13"
-julia = "^1.1"
+julia = "^1.5"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"